n32: fix U32 keyswitch end-to-end — mult/rescale now bit-exact vs OpenFHE
Five stacked root causes, each masking the next:
1. Shoup width mismatch (Conv.cu): ModDown2/DecompAndModUpConv U32 branches
   promoted operands to uint64_t, pairing Shoup_mult_64 with the 2^32-scaled
   constants -> quotient underflow, unreduced products, CRT-poisoned base
   conversion. Fixed by running those multiplies in the 32-bit overload.
2. U32 own-alloc limbs were N elements (4N bytes) while the U32 NTT sweeps
   byte-parity tiles: dense aux-less DIGIT limbs got pool-neighbor clobbers.
   VectorGPU grew an alloc field; Limb own-alloc pads U32 v/aux to 2N.
3. Host NTT grid geometry: kernel U32 M=8 (required by the int4 lane code)
   vs 29 host launch sites' M=4 -> grid 64 instead of 32, wrong transposed
   strides/twiddles. Host M is now width-aware (uniform-U32 chain -> 8).
4. U32 INTT transposed store used >>2, collapsing the (j&2) lane offset:
   half the outputs never written, half write-raced. Now >>1 (matches the
   NTT-side load pairing).
5. U32 forward-NTT phase-1 shared distribution had the row selector outside
   A(): (int4*)A(2*(j&2)) + t offset within a row instead of row 2*(j&2)+t.

Plus: negacyclic scales enabled for U32 (were sizeof(T)==8-gated; helpers
are T-generic), width fixes in hoistedRotateDotKSK_2_, addScaleB_/scaleByP_
(empty U32 branches), binomialSquareExtend_ (missing *P), and
add/sub_scale_p_reuse_b___ (spurious extra load*P).

Validated: GPU EvalMult+relinearize and the rescaled second-level mult are
BIT-IDENTICAL to CPU OpenFHE on identical ciphertexts; forward/inverse NTT
verified 65536/65536 against an offline replica of OpenFHE's negacyclic FTT;
hoisted-rotation dot+automorph and moddown chain verified exact per position.
Residual rotation noise tail equals pure-CPU OpenFHE EvalRotate on the same
chain (upstream n32 parameter-regime property, not a GPU defect).

Debug harness (FHE_KS_TRACE / FHE_KS_TRACE_FULL stage dumps) is runtime-gated
and off by default; hot-path unconditional diagnostics removed.

+++ src/CKKS/Conv.cu
@@ -56,9 +56,15 @@ __global__ void ModDown2(void** __restrict__ a, const __grid_constant__ int n, v
                     buff[tid + blockDim.x * i] = modmult<algo>(((uint64_t*)(b[i]))[idx], G_->ModDown_pre_scale[primeid],
                                                                primeid, G_->ModDown_pre_scale_shoup[primeid]);
                 } else {
+                    // U32 primes carry 2^32-scaled Shoup constants: the multiply must run in the
+                    // 32-bit overload. Promoting to uint64_t pairs Shoup_mult_64 with a 2^32-scaled
+                    // psi, whose __umul64hi quotient is always 0 -> buff holds the UNREDUCED product
+                    // (~2^56). Still congruent mod p (looks in-range after the final modreduce), but
+                    // base conversion needs the exact representative -> k*p excess -> the converted
+                    // limbs go mutually CRT-inconsistent.
                     buff[tid + blockDim.x * i] =
-                        modmult<algo>((uint64_t)((uint32_t*)b[i])[idx], G_->ModDown_pre_scale[primeid], primeid,
-                                      G_->ModDown_pre_scale_shoup[primeid]);
+                        modmult<algo>(((uint32_t*)b[i])[idx], (uint32_t)G_->ModDown_pre_scale[primeid], primeid,
+                                      (uint32_t)G_->ModDown_pre_scale_shoup[primeid]);
                 }
             }
             /*
@@ -196,10 +202,13 @@ __global__ void DecompAndModUpConv(void** __restrict__ a, const int __grid_const
                     ((uint64_t*)(a[pos]))[idx], G_->DecompAndModUp_pre_scale[MODUPIDX_SCALE(d, n_d_n - 1, primeid)],
                     primeid, G_->DecompAndModUp_pre_scale_shoup[MODUPIDX_SCALE(d, n_d_n - 1, primeid)]);
             } else {
+                // See ModDown2: 2^32-scaled Shoup constants require the 32-bit multiply; the
+                // uint64_t promotion left buff unreduced (~2^56) and poisoned the digit base
+                // conversion with k*p multiples.
                 buff[tid + blockDim.x * i_] =
-                    modmult<algo>((uint64_t)((uint32_t*)a[pos])[idx],
-                                  G_->DecompAndModUp_pre_scale[MODUPIDX_SCALE(d, n_d_n - 1, primeid)], primeid,
-                                  G_->DecompAndModUp_pre_scale_shoup[MODUPIDX_SCALE(d, n_d_n - 1, primeid)]);
+                    modmult<algo>(((uint32_t*)a[pos])[idx],
+                                  (uint32_t)G_->DecompAndModUp_pre_scale[MODUPIDX_SCALE(d, n_d_n - 1, primeid)], primeid,
+                                  (uint32_t)G_->DecompAndModUp_pre_scale_shoup[MODUPIDX_SCALE(d, n_d_n - 1, primeid)]);
             }
         }
         /*
